+++ chrome/content.js
@@ -1,15 +1,11 @@
-chrome.runtime.onMessage.addListener(function (message) {
-    if (message.__autorun__ !== undefined) {
-        window.postMessage(message, '*');
-    }
-});
+chrome.runtime.onMessage.addListener(message =>
+    message.__autorun__ === undefined
+        ? undefined
+        : window.postMessage(message, '*')
+);
 
-window.addEventListener('message', function (event) {
-    if (event.source === window && event.data.__timestamp__) {
-        chrome.runtime.sendMessage(event.data, function (requestStop) {
-            if (requestStop) {
-                window.postMessage({ __autorun__: false }, '*');
-            }
-        });
-    }
-}, false);
+window.addEventListener('message', event =>
+    event.source === window && event.data.__timestamp__
+        ? chrome.runtime.sendMessage(event.data, requestStop => requestStop && window.postMessage({ __autorun__: false }, '*'))
+        : undefined
+);