+++ chrome/devtools.js
@@ -1,1 +1,1 @@
-chrome.devtools.panels.create('MiniMongoExplorer', 'images/icon64.png', 'index.html');
+chrome.devtools.panels.create('MiniMongoExplorer', 'images/icon64.png', 'panel.html');