+++ extension/translations/en.js
@@ -1,0 +1,28 @@
+export default {
+    help: {
+        toggle: 'Toggle help'
+    },
+
+    reactivity: {
+        disable: 'Disable reactivity',
+        enable:  'Enable reactivity'
+    },
+
+    sidebar: {
+        hide: 'Hide sidebar',
+        show: 'Show sidebar'
+    },
+
+    text: {
+        disable: 'Disable text mode',
+        enable:  'Enable text mode'
+    },
+
+    ui: {
+        close:       'Close all tabs',
+        collections: 'Collections',
+        query:       'Query',
+        refresh:     'Refresh snapshot',
+        sort:        'Sort'
+    }
+};